+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#0.5.1
+The Zipkin Rack middleware will not raise an error if sending information to Zipkin raises an error
+Integration specs to make sure information is properly passed when using Rack middleware together with Faraday's
+
 # 0.5.0
 Added Faraday middleware to the repo